refactor: extract TCPMessageProtocol and FileDownloadCompletedEventArgs to own files
Split FeedbackTCPClient.cs (was 600+ lines with 3 classes crammed in)
into separate files following single-responsibility principle.
